+++ assets/studio/js/src/modules/object-merger/object-merger-page/components/object-merger-view/constants.tsx
@@ -12,3 +12,5 @@ export enum ComparisonCategoryName {
   SYSTEM_DATA = 'systemData',
   META = 'meta'
 }
+
+export const MERGE_SOURCES = ['main', 'target']
